[V0.1.2] Change in the Data sent
In the previous version we sent data without preprocessing and used the arduino to map values to a correct range.

In this version the mapping is done directly in the app before sending.
In the Arduino we now get directly the correct values, reducing the processing power and allowing it to be used for other things.